requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
h11==0.14.0
idna==3.8
openai==1.50.1
outcome==1.3.0.post0
packaging==24.1
psutil==6.0.0
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.24.0
setuptools==74.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlalchemy==2.0.35
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.2.2
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==13.0.1
wsproto==1.2.0
